Refactor UserController to TestController for development purposes - to be deleted later
