+++ src/main/java/com/example/echo_api/controller/test/TestControllerAdvice.java
@@ -1,4 +1,4 @@
-package com.example.echo_api.controller.user;
+package com.example.echo_api.controller.test;
 
 import org.springframework.web.bind.annotation.ControllerAdvice;
 
@@ -6,7 +6,7 @@ import com.example.echo_api.exception.AbstractControllerAdvice;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
-@ControllerAdvice(assignableTypes = UserController.class)
-public class UserControllerAdvice extends AbstractControllerAdvice {
+@ControllerAdvice(assignableTypes = TestController.class)
+public class TestControllerAdvice extends AbstractControllerAdvice {
 
 }
