+++ src/test/java/com/example/echo_api/unit/controller/TestControllerAdvice.java
@@ -18,17 +18,17 @@ import org.springframework.test.web.servlet.MockMvc;
 
 import com.example.echo_api.config.ApiConfig;
 import com.example.echo_api.config.ErrorMessageConfig;
-import com.example.echo_api.controller.user.UserController;
+import com.example.echo_api.controller.test.TestController;
 import com.example.echo_api.exception.custom.username.UsernameNotFoundException;
 import com.example.echo_api.persistence.model.User;
 import com.example.echo_api.service.account.AccountService;
 
 /**
- * Unit test class for {@link UserController}.
+ * Unit test class for {@link TestController}.
  */
-@WebMvcTest(UserController.class)
+@WebMvcTest(TestController.class)
 @AutoConfigureMockMvc(addFilters = false)
-class UserControllerTest {
+class TestControllerAdvice {
 
     @Autowired
     private MockMvc mockMvc;
@@ -51,7 +51,7 @@ class UserControllerTest {
     @Test
     void UserController_FindAll_ReturnListOfUser() throws Exception {
         // api: GET /api/v1/user/list ==> 200 : List<User>
-        String endpoint = ApiConfig.User.FIND_ALL;
+        String endpoint = ApiConfig.Test.FIND_ALL;
 
         when(accountService.findAll())
             .thenReturn(List.of(testUser));
@@ -66,7 +66,7 @@ class UserControllerTest {
     @Test
     void UserController_FindAll_ReturnListOfEmpty() throws Exception {
         // api: GET /api/v1/user/list ==> 200 : No Content
-        String endpoint = ApiConfig.User.FIND_ALL;
+        String endpoint = ApiConfig.Test.FIND_ALL;
 
         when(accountService.findAll())
             .thenReturn(List.of());
@@ -81,7 +81,7 @@ class UserControllerTest {
     @Test
     void UserController_FindByUsername_ReturnUser() throws Exception {
         // api: GET /api/v1/user/list/{username} ==> 200 : User
-        String endpoint = ApiConfig.User.FIND_ALL + "/" + testUser.getUsername();
+        String endpoint = ApiConfig.Test.FIND_ALL + "/" + testUser.getUsername();
 
         when(accountService.findByUsername(testUser.getUsername()))
             .thenReturn(testUser);
@@ -97,7 +97,7 @@ class UserControllerTest {
     @Test
     void UserController_FindByUsername_Return400UsernameNotFound() throws Exception {
         // api: GET /api/v1/user/list/{username} ==> 400 : Username Not Found
-        String endpoint = ApiConfig.User.FIND_ALL + "/" + testUser.getUsername();
+        String endpoint = ApiConfig.Test.FIND_ALL + "/" + testUser.getUsername();
 
         when(accountService.findByUsername(testUser.getUsername()))
             .thenThrow(new UsernameNotFoundException());
